+++ plugins/cube/shaders-3-2.tpp
@@ -157,7 +157,7 @@ R"(#version 320 es
 
 in highp vec2 guv;
 in highp vec3 colorFactor;
-layout(location = 0) out mediump vec4 outColor;
+layout(location = 0) out highp vec4 outColor;
 
 uniform sampler2D smp;
