+++ src/core/shaders.tpp
@@ -1,7 +1,7 @@
 static const char *default_vertex_shader_source =
 R"(#version 100
 
-attribute mediump vec2 position;
+attribute highp vec2 position;
 attribute highp vec2 uvPosition;
 varying highp vec2 uvpos;
 
@@ -18,11 +18,11 @@ R"(#version 100
 @builtin@
 
 varying highp vec2 uvpos;
-uniform mediump vec4 color;
+uniform highp vec4 color;
 
 void main()
 {
-    mediump vec4 tex_color = get_pixel(uvpos);
+    highp vec4 tex_color = get_pixel(uvpos);
     tex_color.rgb = tex_color.rgb * color.a;
     gl_FragColor = tex_color * color;
 })";
@@ -30,7 +30,7 @@ void main()
 static const char *color_rect_fragment_source =
 R"(#version 100
 varying highp vec2 uvpos;
-uniform mediump vec4 color;
+uniform highp vec4 color;
 
 void main()
 {
@@ -42,10 +42,10 @@ void main()
 static const char *builtin_rgba_source =
 R"(
 uniform sampler2D _wayfire_texture;
-uniform mediump vec2 _wayfire_uv_base;
-uniform mediump vec2 _wayfire_uv_scale;
+uniform highp vec2 _wayfire_uv_base;
+uniform highp vec2 _wayfire_uv_scale;
 
-mediump vec4 get_pixel(highp vec2 uv) {
+highp vec4 get_pixel(highp vec2 uv) {
     uv = _wayfire_uv_base + _wayfire_uv_scale * uv;
     return texture2D(_wayfire_texture, uv);
 }
@@ -54,10 +54,10 @@ mediump vec4 get_pixel(highp vec2 uv) {
 static const char *builtin_rgbx_source =
 R"(
 uniform sampler2D _wayfire_texture;
-uniform mediump vec2 _wayfire_uv_base;
-uniform mediump vec2 _wayfire_uv_scale;
+uniform highp vec2 _wayfire_uv_base;
+uniform highp vec2 _wayfire_uv_scale;
 
-mediump vec4 get_pixel(highp vec2 uv) {
+highp vec4 get_pixel(highp vec2 uv) {
     uv = _wayfire_uv_base + _wayfire_uv_scale * uv;
     return vec4(texture2D(_wayfire_texture, uv).rgb, 1.0);
 }
@@ -66,10 +66,10 @@ mediump vec4 get_pixel(highp vec2 uv) {
 static const char *builtin_external_source =
 R"(
 uniform samplerExternalOES _wayfire_texture;
-uniform mediump vec2 _wayfire_uv_base;
-uniform mediump vec2 _wayfire_uv_scale;
+uniform highp vec2 _wayfire_uv_base;
+uniform highp vec2 _wayfire_uv_scale;
 
-mediump vec4 get_pixel(highp vec2 uv) {
+highp vec4 get_pixel(highp vec2 uv) {
     uv = _wayfire_uv_base + _wayfire_uv_scale * uv;
     return texture2D(_wayfire_texture, uv);
 }
